create a separate namespace for classes shared between debug server and vsix
