+++ VSRAD.DebugServer/Handlers/FetchResultRangeHandler.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using System.Threading.Tasks;
 using VSRAD.DebugServer.IPC.Responses;
+using VSRAD.DebugServer.SharedUtils;
 
 namespace VSRAD.DebugServer.Handlers
 {
