+++ VSRAD.DebugServer/Handlers/ListFilesHandler.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using VSRAD.DebugServer.IPC.Commands;
 using VSRAD.DebugServer.IPC.Responses;
+using VSRAD.DebugServer.SharedUtils;
 
 namespace VSRAD.DebugServer.Handlers
 {
