+++ VSRAD.DebugServer/SharedUtils/ObservableProcess.cs
@@ -4,7 +4,7 @@ using System.Text;
 using System.Threading.Tasks;
 using VSRAD.DebugServer.IPC.Responses;
 
-namespace VSRAD.DebugServer
+namespace VSRAD.DebugServer.SharedUtils
 {
     public sealed class ObservableProcess
     {
